[2/2] Rename RepositoryService to IndexService
Rename module
